backend/requirements.txt
Flask>=3,<4
Flask-Cors>=3,<4
PyJWT
